fix(Trader): Keep trading after exception
If we attempt to trade invalid trades, a.k.a trades where
the whitelist have allowed us to trade with bad tokens
then our local node seams to test the trade and not propagate
it on revert. In that case we want to add the bad token to the
blacklist and keep trading.